Install platform addons only when installing on platform
When we run the installation on the platform, we have several mandatory
environment variables used by the cloud platform (such as
attachment_s3). AWS_ACCESS_KEY_ID is one of them, and we can assume that
if this key is defined, we are installing the instance on your platform.

In that very case, we install and setup the cloud platform addons.

In the image delivered to BSO, those operations will just be skipped.

BSBSO10-10